lab: PixelPanel and DirectorPanel are compositions
PixelPanel.tsx was a 957-line component; it is 240 lines that compose one hook per concern under panels/pixel/ - tool state, the select tool, the viewport, painting the three canvases, pointer gestures, the keyboard - and two toolbar components. DirectorPanel.tsx was 897; it is 230 over panels/director/ - the transport view, time zoom and lane height, show editing with its drags, the show as files, and the two bars. Every body moved verbatim where it could, and the 23 tests from the previous commit pass unchanged. Two things the move taught: refs stay in the component that renders the element and are handed to hooks, because the React Compiler rule refuses a ref read through a hook's return during render; and a helper module cannot be called layout.ts anywhere under app/, because Next reads that name as a route layout.
